--- pychan/decorators/rate_me.py
@@ -1,5 +1,0 @@
-class RateMe:
-    @classmethod
-    async def rate_me(cls, ctx):
-        """NOT FINISHED """
-        pass
